+++ integration/auth/auth_test.go
@@ -58,7 +58,10 @@ func TestMain(m *testing.M) {
 	if err != nil {
 		log.Fatalln(err)
 	}
-	rand.Seed(time.Now().UTC().UnixNano())
+
+	seed := time.Now().UTC().UnixNano()
+	log.Printf("Using random seed: %d", seed)
+	rand.Seed(seed)
 	os.Exit(m.Run())
 }